rename: 요청과 응답에 대한 패키지 분리
