+++ backend/ddang/src/main/java/com/ddang/ddang/bid/application/dto/CreateBidDto.java
@@ -3,7 +3,7 @@ package com.ddang.ddang.bid.application.dto;
 import com.ddang.ddang.auction.domain.Auction;
 import com.ddang.ddang.auction.domain.Price;
 import com.ddang.ddang.bid.domain.Bid;
-import com.ddang.ddang.bid.presentation.dto.CreateBidRequest;
+import com.ddang.ddang.bid.presentation.dto.request.CreateBidRequest;
 import com.ddang.ddang.user.domain.User;
 
 public record CreateBidDto(Long auctionId, int price) {
