+++ backend/ddang/src/main/java/com/ddang/ddang/bid/presentation/BidController.java
@@ -4,10 +4,10 @@ import com.ddang.ddang.bid.application.BidService;
 import com.ddang.ddang.bid.application.dto.CreateBidDto;
 import com.ddang.ddang.bid.application.dto.LoginUserDto;
 import com.ddang.ddang.bid.application.dto.ReadBidDto;
-import com.ddang.ddang.bid.presentation.dto.CreateBidRequest;
-import com.ddang.ddang.bid.presentation.dto.LoginUserRequest;
-import com.ddang.ddang.bid.presentation.dto.ReadBidResponse;
-import com.ddang.ddang.bid.presentation.dto.ReadBidsResponse;
+import com.ddang.ddang.bid.presentation.dto.request.CreateBidRequest;
+import com.ddang.ddang.bid.presentation.dto.request.LoginUserRequest;
+import com.ddang.ddang.bid.presentation.dto.response.ReadBidResponse;
+import com.ddang.ddang.bid.presentation.dto.response.ReadBidsResponse;
 import com.ddang.ddang.bid.presentation.resolver.LoginUser;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
