--- backend/ddang/src/main/java/com/ddang/ddang/bid/presentation/dto/LoginUserRequest.java
@@ -1,4 +1,0 @@
-package com.ddang.ddang.bid.presentation.dto;
-
-public record LoginUserRequest(Long userId) {
-}
