+++ backend/ddang/src/main/java/com/ddang/ddang/bid/presentation/dto/request/CreateBidRequest.java
@@ -1,4 +1,4 @@
-package com.ddang.ddang.bid.presentation.dto;
+package com.ddang.ddang.bid.presentation.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
