+++ backend/ddang/src/main/java/com/ddang/ddang/bid/presentation/dto/request/LoginUserRequest.java
@@ -1,0 +1,4 @@
+package com.ddang.ddang.bid.presentation.dto.request;
+
+public record LoginUserRequest(Long userId) {
+}
